Fix crawl_page function for old parsers
The old, monolithic crawler inherits from BaseCrawler, but it overrides
the initialize function. It calls each state on a loop, setting several
instance variables as it does so.

The new crawlers set most of these variables on initialization. The new
crawlers use a new results intance variable, which the crawl_page method
expects to exist and have a certain format.

To keep the old crawlers working, I've extracted the initial results
hash to its own method, which can then be called by the new initialize
function and inside the old loop for the old parsers.